// src/api/generated/models/authMeRoleDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Bookora API
 * Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.
 * OpenAPI spec version: 1.0
 */
import type { AuthMeRoleDtoType } from './authMeRoleDtoType';

export interface AuthMeRoleDto {
  /**
     * @minLength 26
     * @maxLength 26
     * @pattern ^[0-7][0-9A-HJKMNP-TV-Z]{25}$
     */
  id: string;
  code: string;
  level: number;
  type: AuthMeRoleDtoType;
  isSystem: boolean;
}